requirements.txt
asgiref==3.5.0
boto3==1.21.22
botocore==1.24.22
click==8.0.4
colorama==0.4.4
Deprecated==1.2.13
Django==3.2.11
django-admin-honeypot==1.1.0
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-js-asset==1.2.2
django-redis==5.0.0
django-robots==4.0
django-rq==2.4.1
django-storages==1.11.1
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-jwt==1.0.1
djangorestframework-simplejwt==5.0.0
jmespath==1.0.0
mysqlclient==2.1.0
packaging==21.3
Pillow==9.0.1
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
redis==4.1.4
rq==1.10.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
wrapt==1.14.0
